# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OMNI-TITANIUM 7.0 – Forensic Engine API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /parlays:
    post:
      operationId: saveParlay
      tags: [predictions]
      summary: Save a parlay to history
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveParlayInput"
      responses:
        "200":
          description: Saved parlay
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveParlayOut"

    get:
      operationId: listParlays
      tags: [predictions]
      summary: List saved parlays (history)
      responses:
        "200":
          description: Parlay history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParlayHistoryItem"
  /parlays/resolve:
    post:
      operationId: resolveParlay
      tags: [predictions]
      summary: Mark a parlay as hit/near-miss/miss
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveParlayInput"
      responses:
        "200":
          description: Resolution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveParlayOut"
  /parlays/{parlayId}:
    delete:
      operationId: deleteParlay
      tags: [predictions]
      summary: Delete a saved parlay
      parameters:
        - name: parlayId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
...
components:
  schemas:
...
    SaveParlayInput:
      type: object
      properties:
        date: { type: string }
        label: { type: string }
        legsJson: { type: string }
        jointProb: { type: number }
        combinedFairOdd: { type: number }
        confidenceLabel: { type: string }
        totalLegs: { type: integer }
      required: [date, label, legsJson, jointProb, combinedFairOdd, confidenceLabel, totalLegs]
    SaveParlayOut:
      type: object
      properties:
        success: { type: boolean }
        parlayId: { type: string }
        id: { type: integer }
      required: [success, parlayId]
    ParlayHistoryItem:
      type: object
      properties:
        id: { type: integer }
        parlayId: { type: string }
        date: { type: string }
        label: { type: string }
        generatedAt: { type: string }
        legsJson: { type: string }
        jointProb: { type: number }
        combinedFairOdd: { type: number }
        confidenceLabel: { type: string }
        totalLegs: { type: integer }
        status: { type: string }
        resolvedAt: { type: string, nullable: true }
        hitLegs: { type: integer, nullable: true }
        nearMissLegs: { type: integer, nullable: true }
        resultLabel: { type: string, nullable: true }
        actualResultsJson: { type: string, nullable: true }
      required: [id, parlayId, date, label, generatedAt, legsJson, jointProb, combinedFairOdd, confidenceLabel, totalLegs, status]
    ResolveParlayInput:
      type: object
      properties:
        parlayId: { type: string }
        hitLegs: { type: integer }
        nearMissLegs: { type: integer }
        actualResultsJson: { type: string, nullable: true }
      required: [parlayId, hitLegs, nearMissLegs]
    ResolveParlayOut:
      type: object
      properties:
        success: { type: boolean }
        parlayId: { type: string }
        status: { type: string }
        resultLabel: { type: string }
        hitLegs: { type: integer }
        nearMissLegs: { type: integer }
      required: [success, parlayId, status, resultLabel, hitLegs, nearMissLegs]
...