# lib/api-zod/src/generated/types/parlayHistoryItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OMNI-TITANIUM 7.0 – Forensic Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface ParlayHistoryItem {
  id: number;
  parlayId: string;
  date: string;
  label: string;
  generatedAt: string;
  legsJson: string;
  jointProb: number;
  combinedFairOdd: number;
  confidenceLabel: string;
  totalLegs: number;
  status: string;
  resolvedAt?: string | null;
  hitLegs?: number | null;
  nearMissLegs?: number | null;
  resultLabel?: string | null;
  actualResultsJson?: string | null;
}