# lib/api-zod/src/generated/types/resolveParlayInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OMNI-TITANIUM 7.0 – Forensic Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface ResolveParlayInput {
  parlayId: string;
  hitLegs: number;
  nearMissLegs: number;
  actualResultsJson?: string | null;
}