# lib/api-zod/src/generated/types/resolveParlayOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OMNI-TITANIUM 7.0 – Forensic Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface ResolveParlayOut {
  success: boolean;
  parlayId: string;
  status: string;
  resultLabel: string;
  hitLegs: number;
  nearMissLegs: number;
}