# lib/api-zod/src/generated/types/saveParlayInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OMNI-TITANIUM 7.0 – Forensic Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveParlayInput {
  date: string;
  label: string;
  legsJson: string;
  jointProb: number;
  combinedFairOdd: number;
  confidenceLabel: string;
  totalLegs: number;
}