# lib/api-zod/src/generated/types/saveParlayOut.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OMNI-TITANIUM 7.0 – Forensic Engine API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveParlayOut {
  success: boolean;
  parlayId: string;
  id?: number;
}